add SettingsUiState
This commit introduces the `SettingsUiState` sealed interface to represent the different states of the settings screen: `Loading` and `Success`.
The `Success` state holds the current UI theme and a boolean indicating whether dynamic colors are used.

Additionally, a dependency on `projects.core.data` is added to the `feature/settings` module.