feat(alarm): add AlarmReceiver
This commit introduces the `AlarmReceiver` class.

The `AlarmReceiver` is a `BroadcastReceiver` that is responsible for handling alarms. When an alarm is triggered, the `onReceive` method is called. This method retrieves the ID of the alarm from the intent extras and then uses the `Notifier` to post a reminder notification for that alarm.

The `AlarmReceiver` is annotated with `@AndroidEntryPoint` to enable Hilt dependency injection.